refactor: rename folder to utils
